implemented button for setup and renamed widgets directory to shared
