+++ lib/shared/static_navigation_bar.dart
