+++ lib/shared/widgets.dart
@@ -1,5 +1,6 @@
 export 'dynamic_text.dart';
 export 'loading.dart';
+export 'register_button.dart';
 export 'sliver_loading_indicator.dart';
 export 'static_navigation_bar.dart';
 export 'tag.dart';
