+++ lib/views/cafeteria/cafeteria.dart
@@ -3,7 +3,7 @@ import 'package:provider/provider.dart';
 
 import '../../models/models.dart';
 import '../../services/services.dart';
-import '../../widgets/widgets.dart';
+import '../../shared/widgets.dart';
 import 'local_widgets/local_widgets.dart';
 
 class Cafeteria extends StatelessWidget {
