+++ lib/views/calendar/calendar_event/calendar_event.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../../../models/models.dart';
-import '../../../widgets/widgets.dart';
+import '../../../shared/widgets.dart';
 import 'local_widgets/local_widgets.dart';
 
 class CalendarEvent extends StatelessWidget {
