+++ lib/views/news/local_widgets/news_list_panel.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
 import '../../../models/models.dart';
+import '../../../shared/widgets.dart';
 import '../../../styles/styles.dart';
 import '../../../utils/utils.dart';
-import '../../../widgets/widgets.dart';
 import '../news_article/news_article.dart';
 
 class NewsListPanel extends StatelessWidget {
