+++ lib/views/news/news_article/local_widgets/news_article_panel.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
 import '../../../../models/models.dart';
+import '../../../../shared/widgets.dart';
 import '../../../../styles/styles.dart';
-import '../../../../widgets/widgets.dart';
 
 class NewsArticlePanel extends StatelessWidget {
   final Article article;
