+++ lib/views/organisation/organisation_examregulations/organisation_examregulations.dart
@@ -2,8 +2,8 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 
+import '../../../shared/widgets.dart';
 import '../../../styles/styles.dart';
-import '../../../widgets/widgets.dart';
 
 class VerwaltungPruefungsOrdnung extends StatefulWidget {
   @override
