+++ lib/views/setup/sign_in.dart
@@ -1,7 +1,7 @@
 // import 'package:flutter/cupertino.dart';
 // import 'package:flutter/material.dart';
 // import 'package:mateapp/services/services.dart';
-// import 'package:mateapp/widgets/widgets.dart';
+// import 'package:mateapp/shared/widgets.dart';
 // import 'package:mateapp/views/views.dart';
 // import 'package:mateapp/models/models.dart';
